consolidate RPL commands
